Small improvements for token prices to show on testnet
