+++ .env.example
@@ -6,6 +6,10 @@ REACT_APP_GOOGLE_ANALYTICS_ID=
 REACT_APP_BALANCER_POOL_URL=
 REACT_APP_BALANCER_SUBGRAPH_URL=
 
+REACT_APP_BALANCER_POOLS_ADDRESSES=0x0,0x0
+REACT_APP_UNI_POOLS_ADDRESSES=0x0,0x0
+REACT_APP_SUSHI_POOLS_ADDRESSES=0x0,0x0
+
 REACT_APP_HALO_TOKEN_ADDRESS_MAINNET=
 REACT_APP_HALO_REWARDS_ADDRESS_MAINNET=
 REACT_APP_HALOHALO_ADDRESS_MAINNET=
